Now every subsequent request will have token from locaStorage which gets stored upon signUp or logIn
Now what i did is that once the users sign in aur login then a token will get saved in the localStorage and now for every subsequent request the token will be provided in the headers. Now there is a middleware function in the server which will verify requests for all the /user route. This will only be accessible when the user is logged in or sign in